refactor(navbar): make navbar component determine which version to use
